requirements.txt
python-dotenv
requests
random
os
logging 
